Lua demuxers: show the original URL in playlist
If playlists containing youtube urls are exported or imported, these patches
improve the usability of the playlist.

Lua remembers the original url (i.e. the website _containing_ the content) if
no other url is passed by the lua parsers. It also puts the content title to
the meta-title entry.

XSPF export / import now uses the <info> to deal with the meta-url.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>