equalizer: Optimize calculations of alpha, beta, and gamma constants
It is easy to see that the computation of the alpha constant is slightly
inefficient, as the root value can be close to 1 for the lower frequencies,
which would result in significant roundoff error if this value is subtracted
from 1.  In this patch, the computation of the alpha, beta, and gamma constants
has been simplified and refactored.

Acked-by: Ilkka Ollakka <ileoo@videolan.org>
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>