skins2: add a function that differentiates dependent and independent layouts.
In skins2, a window can have multiple layouts with two different goals.

First goal, layouts have the same original size and only differ in the
presentation (different color, ...). If the active layout gets resized,
the user expects the related inactive layouts with the same size to
automatically resize, should they become active.

Second goal, layouts are different in size, e.g a reduced layout and an
expanded layout. In this case, resizing the active layout doesn't mean
resizing the other inactive layouts, since they were not meant to be
similar in size in the first place.

This patch creates a function that will be used to differentiate these
two different use of layouts.